+++ include/seqan3/range/views/kmer_hash.hpp
@@ -223,6 +223,9 @@ private:
     //!\brief The sentinel type of the underlying range.
     using sentinel_t = std::ranges::sentinel_t<rng_t>;
 
+    template <typename urng2_t>
+    friend class shape_iterator;
+
 public:
     /*!\name Associated types
      * \{
@@ -252,6 +255,19 @@ public:
     constexpr shape_iterator & operator=(shape_iterator const &) = default; //!< Defaulted.
     constexpr shape_iterator & operator=(shape_iterator &&)      = default; //!< Defaulted.
     ~shape_iterator()                                            = default; //!< Defaulted.
+
+    //!\brief Allow iterator on a const range to be constructible from an iterator over a non-const range.
+    template <typename urng2_t>
+    //!\cond
+        requires std::same_as<std::remove_const_t<urng_t>, urng2_t>
+    //!\endcond
+    shape_iterator(shape_iterator<urng2_t> it) :
+        hash_value{std::move(it.hash_value)},
+        roll_factor{std::move(it.roll_factor)},
+        shape_{std::move(it.shape_)},
+        text_left{std::move(it.text_left)},
+        text_right{std::move(it.text_right)}
+    {}
 
     /*!\brief Construct from a given iterator on the text and a seqan3::shape.
     * /param[in] it_start Iterator pointing to the first position of the text.
@@ -491,14 +507,12 @@ public:
     /*!\brief Move the iterator by a given offset and return the corresponding hash value.
      * \attention This function is only avaible if `it_t` models std::random_access_iterator.
      */
-    value_type operator[](difference_type const n)
+    reference operator[](difference_type const n) const
     //!\cond
         requires std::random_access_iterator<it_t>
     //!\endcond
     {
-        text_left += n;
-        hash_full();
-        return operator*();
+        return *(*this + n);
     }
 
     //!\brief Return the hash value.